requirements.txt
Django
django-ckeditor==6.1.0
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow
python-dotenv>=0.19.1
django-simple-captcha==0.5.14
django-crispy-forms==1.13.0
whitenoise==5.3.0
django-countries==7.2.1
django-autoslug==1.9.8
